feat: re-indexing login and form
