+++ next.config.js
@@ -1,10 +1,15 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
+  webpack: (config, { dev, isServer }) => {
+    // Overwrite the default configuration
+    if (!isServer) {
+      // Use 'source-map' for the development environment and 'hidden-source-map' for production
+      config.devtool = dev ? 'source-map' : 'hidden-source-map';
+    }
+
+    return config;
+  },
 }
 
-const withBundleAnalyzer = require('@next/bundle-analyzer')({
-  enabled: process.env.ANALYZE === 'true',
-})
-
-module.exports = withBundleAnalyzer(nextConfig)
+module.exports = nextConfig
