+++ src/pages/api/signup.ts
@@ -17,25 +17,3 @@ const signup = async (req: NextApiRequest, res: NextApiResponse) => {
 };
 
 export default signup;
-
-// import { db } from '@/firebase/firebaseClient';
-// import { doc, setDoc } from 'firebase/firestore';
-// import type { NextApiRequest, NextApiResponse } from 'next';
-
-// const signup = async (req: NextApiRequest, res: NextApiResponse) => {
-//   const { uid, email } = req.body;
-
-//   try {
-//     const user = await setDoc(doc(db, 'users', uid), {
-//       uid: uid,
-//       email: email,
-//       activeSubscription: false,
-//       sessionId: '',
-//     });
-//     res.status(200).json({ userStatus: 'initialized', user });
-//   } catch (error) {
-//     res.status(400).json({ error });
-//   }
-// };
-
-// export default signup;
